requirements.txt
absl-py==0.12.0
astunparse==1.6.3
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
colorlog==5.0.1
EasyProcess==0.3
entrypoint2==0.2.4
gast==0.3.3absl-py==0.11.0
astunparse==1.6.3
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
colorlog==4.8.0
cycler==0.10.0
decorator==4.4.2
EasyProcess==0.3
entrypoint2==0.2.4
flatbuffers==1.12
gast==0.3.3
google-auth==1.27.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
imgaug==0.4.0
imutils==0.5.4
iso-639==0.4.5
iso3166==1.0.1
isodate==0.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.3.4
MouseInfo==0.1.3
mss==6.1.0
networkx==2.5
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
pafy==0.5.5
pandas==1.2.2
Pillow==8.1.0
protobuf==3.15.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.10.1
PyDirectInput==1.0.4
pygame==2.0.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==2.4.7
pyperclip==1.8.2
pypiwin32==223
PyRect==0.1.4
pyscreenshot==2.3
PyScreeze==0.1.26
PySocks==1.7.1
python-dateutil==2.8.1
PyTweening==1.0.3
pytz==2021.1
pyvjoy==1.0.1
PyWavelets==1.1.1
pywin32==300
PyYAML==5.4.1
pyzmq==22.0.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.1
scipy==1.4.1
Shapely==1.7.1
six==1.15.0
streamlink==2.1.1
tensorboard==2.2.2
tensorboard-plugin-wit==1.8.0
tensorflow-gpu @ https://storage.googleapis.com/tensorflow/windows/gpu/tensorflow_gpu-2.2.0-cp38-cp38-win_amd64.whl
tensorflow-gpu-estimator==2.2.0
termcolor==1.1.0
tf-slim==1.1.0
tflearn==0.5.0
tifffile==2021.3.5
tqdm==4.58.0
typing-extensions==3.7.4.3
urllib3==1.26.3
vidgear==0.2.0
websocket-client==0.58.0
Werkzeug==1.0.1
wrapt==1.12.1
youtube-dl==2021.4.1
